% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{koi}
\alias{koi}
\title{Public Domain Komi-Permyak data}
\format{Data frame with a Komi-Permyak book
\describe{
  \item{doc_id}{name of the text in original corpus}
  \item{sentence_id}{sentence id, unique within a text}
  \item{sentence}{sentence text}
  ...
}}
\source{
\url{http://urn.fi/URN:NBN:fi-fe2014101045137}
}
\usage{
koi
}
\description{
A dataset containing sentences from Public Domain sources.
Four Battles book proofread by Niko Partanen.
}
\keyword{datasets}
