% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kpv}
\alias{kpv}
\title{Public Domain Komi-Zyrian data}
\format{Data frame with Komi-Zyrian data
\describe{
  \item{doc_id}{name of the text in original corpus}
  \item{sentence_id}{sentence id, unique within a text}
  \item{sentence}{sentence text}
  ...
}}
\source{
\url{https://github.com/langdoc/kpv-lit}

\url{https://fennougrica.kansalliskirjasto.fi/}

\url{http://komikyv.org/}
}
\usage{
kpv
}
\description{
A dataset containing sentences from Public Domain sources.
Digitalized in Fenno-Ugrica and proofread in FU-Lab.
Ivan Belykh's works made available by the author. Four Battles
book proofread by Niko Partanen. Current version contains only data
comparable with other languages, also the idea is to have all possible
Komi-Zyrian data here and the sentence_id would tell which are
matching.
}
\keyword{datasets}
