% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mhr}
\alias{mhr}
\title{Public Domain Meadow Mari data}
\format{Data frame with a Meadow Mari book
\describe{
  \item{doc_id}{name of the text in original corpus}
  \item{sentence_id}{sentence id, unique within a text}
  \item{sentence}{sentence text}
  ...
}}
\source{
\url{http://urn.fi/URN:NBN:fi-fe2014091844781}
}
\usage{
mhr
}
\description{
A dataset containing sentences from Public Domain sources.
Four Battles book proofread by Niko Partanen.
}
\keyword{datasets}
