% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psdp}
\alias{psdp}
\title{Public Pite Saami data}
\format{Example of digitalized Pite Saami data
\describe{
  \item{doc_id}{name of the text in original corpus}
  \item{sentence_id}{sentence id, unique within a text}
  \item{sentence}{sentence text}
  \item{sentence_eng}{English translation}
  \item{sentence_hun}{Hungarian translation}
  \item{sentence_orig_orth}{Original orthography}
  ...
}}
\source{
\url{http://saami.uni-freiburg.de/psdp/syntax}
}
\usage{
psdp
}
\description{
sje18930000-halasz1893a-22
(public domain; digitized, transcribed and annotated
by the Pite Saami Documentation Project (Joshua Wilbur, Iris Perkmann))
}
\keyword{datasets}
