% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kpv}
\alias{kpv}
\title{Public Domain Komi-Zyrian data}
\format{Data frame with 877685 tokens
\describe{
  \item{sentence_id}{sentence id, unique within a text}
  \item{sentence}{sentence text}
  \item{token}{token in lowercase}
  \item{text_name}{name of the text in original corpus}
  ...
}}
\source{
\url{https://github.com/langdoc/kpv-lit}

\url{https://fennougrica.kansalliskirjasto.fi/}

\url{http://komikyv.org/}
}
\usage{
kpv
}
\description{
A dataset containing 877685 tokens from Public Domain sources.
Digitalized in Fenno-Ugrica and proofread in FU-Lab.
Ivan Belykh's works made available by the author.
}
\keyword{datasets}
